:construction: Lift state up. No error, but onClick not working
