+++ src/components/GenreTag/GenreTag.tsx
@@ -1,27 +1,15 @@
 import styles from "./GenreTag.module.css";
 import "../../globals.css";
-import { useEffect, useState } from "react";
 
 type GenreTagProps = {
   text: string;
+  selected: boolean;
+  onClick?: () => void;
 };
 
-function GenreTag({ text }: GenreTagProps): JSX.Element {
-  const [selected, setActive] = useState<boolean>(false);
-
-  useEffect(() => {
-    console.log(selected, text);
-  }, [selected]);
-
-  function handleClick() {
-    selected ? setActive(false) : setActive(true);
-  }
-
+function GenreTag({ text, selected }: GenreTagProps): JSX.Element {
   return (
-    <button
-      className={selected ? styles.tag_selected : styles.tag_unselected}
-      onClick={handleClick}
-    >
+    <button className={selected ? styles.tag_selected : styles.tag_unselected}>
       {text}
     </button>
   );
